+++ server/src/handlers/arena.rs
@@ -13,15 +13,15 @@ pub async fn get(
     fighter: crate::auth::Auth,
     state: &State<EboboState>,
 ) -> Result<Json<Arena>, BadRequest<String>> {
-    let total = Users::find().all(state.db.as_ref()).await.unwrap().len();
-    let queue = Queue::find().all(state.db.as_ref()).await.unwrap().len();
+    let total = Users::find().all(state.db.as_ref()).await.unwrap().len() as i32;
+    let queue = Queue::find().all(state.db.as_ref()).await.unwrap().len() as i32;
 
     let fighter = fighter.fighter.unwrap();
 
     Ok(Json(Arena {
         total,
         queue,
-        rank: fighter.rank,
+        rank: fighter.rank as i32,
         you: fighter.emo,
     }))
 }
